% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{drop_full_na}
\alias{drop_full_na}
\title{Remove all columns that only contain \code{NA} or \code{NaN}.}
\usage{
drop_full_na(data)
}
\arguments{
\item{data}{The \code{dataframe} that needs to be cleaned.}
}
\value{
The cleaned up \code{dataframe}.
}
\description{
It removes columns that only contain \code{NA} to the \code{\link[tidyr:drop_na]{tidyr::drop_na()}} function.
WARNING: \code{\link[=drop_full_na]{drop_full_na()}} will also drop all \code{NaN}.
}
\examples{
```r
 data = data.frame(
     part = c(NA, "test", NaN),
     na = c(NA, NA, NA),
     nan = c(NaN, NaN, NaN),
     str = c("NA", "NaN", NA),
     ...
 );

 drop_full_na(data) # => only without the `na` and `nan` column
```
}
