% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{formatBoolean}
\alias{formatBoolean}
\title{Adjust the data type for each column of the boolean type.}
\usage{
formatBoolean(data)
}
\arguments{
\item{data}{The \code{dataframe} that needs to be cleaned.}
}
\value{
The cleaned up \code{dataframe}.
}
\description{
It dynamically and "automatically" changes the datatype of the provided data based on the format for each column.
}
\details{
More information
Converts all columns to \code{boolean} via the \code{\link[=as.logical]{as.logical()}} function,
that match this regex \verb{^(NA|true|false)$} for each entry.
INFO: Columns containing only \code{NA} or \code{NaN} will be ignored.
}
\examples{
```r
 #pretty self explanatory how to use it
 formatBoolean(data)
```
}
