% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{formatFactor}
\alias{formatFactor}
\title{Adjust the data type for each column of the \code{factor} type.}
\usage{
formatFactor(data)
}
\arguments{
\item{data}{The \code{dataframe} that needs to be cleaned.}
}
\value{
The cleaned up \code{dataframe}.
}
\description{
It dynamically and "automatically" changes the datatype of the provided data based on the format for each column.
}
\details{
More information
Changes the type to a factor if the existing type is \code{character} and the amount of unique items is less than half of all items.
}
\examples{
```r
 #pretty self explanatory how to use it
 formatFactor(data)
```
}
