% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{formatNumber}
\alias{formatNumber}
\title{Adjust the data type for each column of the \code{number} type.}
\usage{
formatNumber(data)
}
\arguments{
\item{data}{The \code{dataframe} that needs to be cleaned.}
}
\value{
The cleaned up \code{dataframe}.
}
\description{
It dynamically and "automatically" changes the datatype of the provided data based on the format for each column.
}
\details{
More information
Changes the type to:
\itemize{
\item \code{numeric} if only \code{NaN}
\item \code{integer} if only full digits exits
\item \code{double} if only \code{integer}s or digits seperated by \code{.} or \verb{,} exists
}
}
\examples{
```r
 #pretty self explanatory how to use it
 formatNumber(data)
```
}
