% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{make.namesCamel}
\alias{make.namesCamel}
\title{Rename columns based on the camelCase standart.}
\usage{
make.namesCamel(names)
}
\arguments{
\item{names}{A \code{collection} of names that needs to be cleaned.}
}
\description{
Changes the list of provided names to be in camelCase. (also removes spaces and other format problems)
}
\details{
More information
It modifies the standart \code{\link[base:make.names]{base::make.names()}} function to the camelCase standart
and can then be used to apply them with the \code{\link[dplyr:rename]{dplyr::rename_with()}} function to a \code{dataframe}.
INFO: The names are first converted to snake snake_case internally via \code{\link[=make.snakeCase]{make.snakeCase()}}.
}
\examples{
```r
 # use rename with to get the colnames and apply the changes
 `rename_with(data, make.namesCamel)`
 #"x_0_test_string_123" => "x0TestString123"
```
}
\seealso{
\code{\link[=make.snakeCase]{make.snakeCase()}} to do the same with snake_case.
}
