% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{make.namesSnake}
\alias{make.namesSnake}
\title{Rename columns based on the snake_case standart.}
\usage{
make.namesSnake(names)
}
\arguments{
\item{names}{A \code{collection} of names that needs to be cleaned.}
}
\value{
The cleaned up \code{collection} of names.
}
\description{
Changes the list of provided names to be in snake_case. (also removes spaces and other format problems)
}
\details{
More information
It modifies the standart \code{\link[base:make.names]{base::make.names()}} function to the snake_case standart
and can then be used to apply them with the \code{\link[dplyr:rename]{dplyr::rename_with()}} function to a \code{dataframe}.
}
\examples{
```r
 # use rename with to get the colnames and apply the changes
 `rename_with(data, make.namesSnake)`
 #"x0TestString123" => "x_0_test_string_123"
```
}
\seealso{
\code{\link[=make.camelCase]{make.camelCase()}} to do the same with camelCase.
}
